Move SparkSupport around.
